health-concierge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "health-concierge"
version = "0.1.0"
description = "Personal health concierge agent via Telegram"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "anthropic>=0.18.0",
    "garminconnect>=0.2.0",
    "stravalib>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
